api/requirements.txt
Django==2.0.2
psycopg2-binary
djangorestframework
djangorestframework-camel-case
